feat: refactor workflows to use Github App instead of PAT (#34)
This PR refactors the GitHub workflow configurations to use GitHub App authentication instead of a PAT and updates documentation in the README accordingly. Key changes include the replacement of legacy workflows (terraform-vcs.yml and terraform-workspace.yml) with new workflows (terraform-stack.yml, terraform-module-test.yml, and terraform-module-release.yml) and adjustments to the codeowners file.